refactor: move Hero and PillarsSection to src/sections
